export task input to its own component
